Thank you for your patience and sorry for the delay. There are two solutions made: First reads from the console and the second one from a .txt file.
All of the necessary information to execute both solutions are in the README.md file.
Said file is located in : src/main/resources/README.md

Again I am sorry for the rushed solutions but as you will read in the README.md
I had a very good reason for it.

Regards!

Rida